requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langgraph>=0.0.40
langsmith>=0.1.0
streamlit>=1.28.0
python-dotenv>=1.0.0
requests>=2.31.0
google-generativeai>=0.3.0
duckduckgo-search>=3.9.0
pydantic>=2.0.0
typing-extensions>=4.0.0
pydantic-core>=2.0.0
langchain-core>=0.1.0
langchain-google-genai>=0.0.11
bs4
plotly
pandas